requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-google-genai==2.0.7
pygithub==2.5.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
